dns解析适配L7-Perf-trait及L7-RRT
**现象及复现步骤**

**原因和解决方案**

**影响范围**

**验证方案**

**涉及分支**

* master

**检查项**

- [ ] 需要同步修改(trident-dump、trident-copy或trident-testsrv)
- [ ] 需要更新依赖
- [ ] 是共性问题(代码中存在类似问题)
- [ ] 打包成功(如果有修改Makefile)

/assign @jinjie

/cc @xiangyang @yuanchao @jinjie

+++ src/common/protocol_logs.rs
@@ -145,7 +145,7 @@ impl From<AppProtoLogsBaseInfo> for flow_log::AppProtoLogsBaseInfo {
 #[derive(Default, Debug, Clone, PartialEq, Eq)]
 pub struct DnsInfo {
     pub trans_id: u16,
-    pub query_type: u16,
+    pub query_type: u8,
     pub query_name: String,
     // 根据查询类型的不同而不同，如：
     // A: ipv4/ipv6地址
